+++ pkg/kql/cast.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/jinzhu/now"
 	"github.com/opencloud-eu/opencloud/pkg/ast"
-	"github.com/opencloud-eu/opencloud/services/search/pkg/query"
 )
 
 func toNode[T ast.Node](in any) (T, error) {
@@ -85,7 +84,7 @@ func toTimeRange(in any) (*time.Time, *time.Time, error) {
 
 	value, err := toString(in)
 	if err != nil {
-		return &from, &to, &query.UnsupportedTimeRangeError{}
+		return &from, &to, &UnsupportedTimeRangeError{}
 	}
 
 	c := &now.Config{
@@ -132,7 +131,7 @@ func toTimeRange(in any) (*time.Time, *time.Time, error) {
 	}
 
 	if from.IsZero() || to.IsZero() {
-		return nil, nil, &query.UnsupportedTimeRangeError{}
+		return nil, nil, &UnsupportedTimeRangeError{}
 	}
 
 	return &from, &to, nil
